# Conflicts:
#	server/entity/damage/source.go
#	server/item/item.go
#	server/player/player.go

+++ server/block/beacon.go
@@ -134,7 +134,7 @@ func (b Beacon) obstructed(pos cube.Pos, w *world.World) bool {
 		return false
 	}
 	// Slow obstructed light calculation, if the fast way out didn't suffice.
-	return w.HighestLightBlocker(pos.X(), pos.Z()) > int16(pos[1])
+	return w.HighestLightBlocker(pos.X(), pos.Z()) > pos[1]
 }
 
 // broadcastBeaconEffects determines the entities in range which could receive the beacon's powers, and
